requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pytz>=2024.1
gtts>=2.5.0
google-api-python-client>=2.125.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
elevenlabs>=1.0.0
